--- .github/workflows/rsac-synchronizer.yml
@@ -1,24 +1,0 @@
-name: Synchronize
-
-on:
-  workflow_dispatch:
-    inputs:
-      payload:
-        required: true
-      ref:
-        required: true
-
-jobs:
-  synchronize:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout
-        uses: actions/checkout@b4ffde65f46336ab88eb53be808477a3936bae11 # v4
-        with:
-          repository: ${{ fromJson(inputs.payload).owner }}/${{ fromJson(inputs.payload).repo }}
-          ref: ${{ inputs.ref }}
-
-      - name: Synchronize
-        uses: jill64/rsac-synchronizer@main
-        with:
-          payload: ${{ inputs.payload }}
