+++ src/index.ts
@@ -1,6 +1,9 @@
 import { octoflare } from 'octoflare'
-import { onCreateRepo } from './onCreateRepo.js'
-import { onPush } from './onPush.js'
+import { onCreateRepo } from './trigger/onCreateRepo.js'
+import { onPush } from './trigger/onPush.js'
+import { applyConfig } from './utils/applyConfig.js'
+import { getConfig } from './utils/getComfig.js'
+import { mergeConfig } from './utils/mergeConfig.js'
 
 export default octoflare(async ({ installation, payload }) => {
   const event = onPush(payload) ?? onCreateRepo(payload)
@@ -17,17 +20,65 @@ export default octoflare(async ({ installation, payload }) => {
     })
   }
 
-  const { owner, repo, ref } = event
+  const { owner, repo } = event
+  const octokit = installation.kit
 
-  await installation.startWorkflow({
-    payload: {
-      owner,
-      repo
-    },
-    ref
+  const rootConfig = await getConfig({
+    owner,
+    repo: '.github',
+    octokit
   })
 
-  return new Response('RSaC Synchronize Workflow Dispatched', {
-    status: 202
+  if (repo !== '.github') {
+    const repoConfig = await getConfig({
+      owner,
+      repo,
+      octokit
+    })
+
+    await applyConfig({
+      octokit,
+      owner,
+      repo,
+      config: mergeConfig(rootConfig, repoConfig)
+    })
+
+    return new Response('Complete Synchronize', {
+      status: 200
+    })
+  }
+
+  const { data: repository } = await octokit.rest.repos.get({
+    owner,
+    repo
+  })
+
+  const { data: allRepo } = await (repository.owner.type !== 'Organization'
+    ? octokit.rest.repos.listForUser({
+        username: owner
+      })
+    : octokit.rest.repos.listForOrg({
+        org: owner
+      }))
+
+  const result = allRepo.map(async (repo) => {
+    const repoConfig = await getConfig({
+      owner: repo.owner.login,
+      repo: repo.name,
+      octokit
+    })
+
+    await applyConfig({
+      octokit,
+      owner: repo.owner.login,
+      repo: repo.name,
+      config: mergeConfig(rootConfig, repoConfig)
+    })
+  })
+
+  await Promise.all(result)
+
+  return new Response('Complete Synchronize for All Repo', {
+    status: 200
   })
 })
